feat(prdoc): extract as standalone reusable package with rich summaries and LLM enhancement
- Moved all prdoc logic from montrs-agent to montrs-prdoc as the canonical
  package; agent now thin re-exports (prdoc_embed, prdoc_generator, changelog)
- Added summary.rs: enriched template summary generation from commit
  narratives, public API extraction, per-package diff stats, and
  breaking change descriptions
- Added llm.rs (feature-gated): Groq, OpenAI, and Ollama providers for
  LLM-enhanced summary rewriting
- Added local_llm.rs (feature-gated): local LLM inference via candle
  with Qwen2-0.5B-Instruct
- Added config.rs: PrDocConfig with prdoc.toml support, LLM provider
  configuration via env vars
- Added standalone CLI binary (features: cli): prdoc show, validate,
  generate, changelog generate/bump/verify
- Added reusable GitHub Action (action.yml) for CI/CD pipelines
- Added README.md with install, usage, library, config docs
- Updated Cargo.toml with feature gates: cli, llm, local-llm
