+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
     "packages/agent",
     "packages/montrs",
     "packages/orm",
+    "packages/prdoc",
     "packages/validator",
     "packages/test",
     "packages/utils",
